Add a Hacker News API client
Add the initial interface and implementation for a Hacker News API
client.

Note that I've opted to "program to an interface" & am developing a
structure inspired by "hexagonal architecture". This allows to create
loosely coupled components for our application.

For this reason, I've opted for an object oriented approach using an
es6 class. An alternative would be to create a module with simple
named export functions.

An interesting observation was that the HN API can return a 200
response with a `null` as body when walking backwards from the latest
item. I've chosen to throw a `ItemNotFoundError` when this happens.
This will allow us to handle the exception (unexpected response) at
the application layer