Add a NewsReader class which consumes the API
Design Choices:
===============

1) Using a Generator
--------------------
We use a generator as a data producer (Iterator). The method below
will return return a generator object (wrapped in a promise) that can
be used to control the data producer. Using a generator allows us to
pause & resume the "data production" of new stories.

The generator is an async generator since we're making network
requests in its body.

More info on Generators https://exploringjs.com/es6/ch_generators.html

2) do while
-----------
Note that our generator will only yield (pause and send a result)
when we get a valid story from the API. The API we use is limited,
it returns both invalid story data and other data objects like
comments or polls.

Using a loop inside the generator's body allows us to keep making
network requests until we receive a value that satisfies our demands.

!! ASSUMPTION
Our loop will run until we reach the last possible id. Given that the
type of the id is a number, we assume the last possible id is `1`.
See also https://hacker-news.firebaseio.com/v0/item/0.json
(returns null) This assumption is based on the HN API and doesn't
scale.

After yielding the last Story, the generator will return {value:
story, done: true} to indicate there is no more data to generate

We've opted for a do while since we want to run through at least 1
iteration and it allows us to encapsulate all network calls in a
single try catch.

2) `try catch` inside `do while`
--------------------------------
We're using a try catch statement to catch any errors that might
happen during the execution of our generator. When a
ItemNotFoundError happens, we don't want to do anything (except maybe
log it). We want to move on and get the next story. Other are thrown
up to the caller of the generator

The try catch is located INSIDE the loop, this ensures that the loop
does not break when an error is thrown.