Connect UI with the business logic
Design Choices
==============

App is a function component so we're able to use React hooks. We use a
useEffect hook with an empty dependency array to ensure the effect
only happens once.

We load an initial batch of stories using the generator method
`reader.getPreviousStory()`. We use a loop to keep activating the
generator until we've reached the inital amount of stories we want to
load. Each time the generator produces a story we update our state so
that the UI will be updated.

When the generator is unable to produce further data we break out of
the loop.

When an error happens we show a message to the user. This can be
viewed by blocking the domain we make the requests to in our
browser's devtools.

The loadStories method can later serve to load more stories on scroll.