Add feature: loadMoreOnScroll
New
===
* Show a loading indicator when stories are being loaded
* Load more stories when the user scrolls to the bottom

Changed
=======
* Wrap `loadStories` with `useCallback` since it's now a dependency
of our second hook

Design Choices
==============
We want to load more stories when the user scrolls to the bottom of
the page. Rather than waiting until the user has reached the bottom,
we will start loading more stories when the user reaches a certain
distance (SCROLL_OFFSET) from the bottom.

The scroll listener is throttled to ensure that the function only gets
invoked max once per every THROTTLE_WAIT_MS milliseconds.

We bail early if the app is already loading stories.

Given that it can take a while for an application to receive new
stories, we immediately load a number (STORIES_TO_LOAD_ON_SCROLL) of
additional stories.

TODO (idea) rather than a fixed offset, we could calculate it based on
the height of the stories. Trying to ensure there are always a
certain amount of stories "below the fold"

TODO add tests

New
===
* Show a loading indicator when stories are being loaded
* Load more stories when the user scrolls to the bottom

Changed
=======