Add offline functionality
Design Choices
==============

It appears the service worker provided with Create React App (CRA)
isn't easy to extend. CRA uses workbox under the hood and doesn't
allow for much customisation without ejecting and/or installing other
packages.

We've opted to handcraft a Service Worker rather than the earlier
mentioned solutions since too much time would be spent on framework
setup.

To be able to include this custom service worker, we had to

* add a serviceworker.js file in `public` with our code.
* write the serviceworker in JavaScript rather than TypeScript since
  we're unable to modify the webpack setup

Note that the Service Worker provided by CRA is still included in the
build. It has been "unregistered" and won't be active but it seems
impossible to remove without "ejecting" from CRA.

Improvements to be made:

* Remove CRA dependency so that we can remove the included service
  worker
* When ejected, convert the custom service worker to TypeScript
* Add the ability to cache assets generated by webpack
* Expire the hacker news content after a certain time

Caching Strategies
------------------

We use 2 different caching strategies. Requests to get the latest item
will use a Network First approach since this resource will update
frequently. Requests to get items will use a Cache First approach
since these do not update often.

Note that this means that items that are updated (deleted/dead) after
we first see them could be out of sync with the API.